wip: child-create: Prototypical IKE_AUX exchange for QSKE mechanism
The CREATE_CHILD_SA exchange is followed by an IKE_AUX exchange, the
CHILD_SA is only installed then (i.e. the SA is not first installed with
regular keys and then replaced when QSKE is complete).

wip: For some errors a more specific notify might be preferable (e.g.
INVALID_SYNTAX if a QSKE payload is missing or an exchange other than
IKE_AUX follows CREATE_CHILD_SA if a QSKE mechanism was negotiated).
wip: This construct would theoretically allow us to use IKE_AUX after
IKE_AUTH to have stronger PFS directly for the first SA (although it
couldn't be combined with classic DH there, as is possible later and for
rekeyings). So maybe relying on childless SAs is the better approach to
have independent keys for CHILD_SAs.